Fixed longs

+++ Graphs/Dijkstra's.cpp
@@ -4,13 +4,13 @@
 * to every other node in O(N*log(N))
 */
 
-void dijkst(int _src, long _dist[]){
+void dijkst(int _src, int _dist[]){
 	for(int i=1; i<=n; ++i) _dist[i] = INFLL;
 	_dist[_src] = 0;
-	set<pair<long,int>> _sss;
+	set<pair<int,int>> _sss;
 	for(int i=1; i<=n; ++i) _sss.insert({_dist[i], i});
 	while(!empty(_sss)){
-		long d; int u;
+		int d; int u;
 		tie(d, u) = *begin(_sss);
 		_sss.erase(begin(_sss));
 		for(auto [v, w]: adj[u]) if(_dist[v] > _dist[u] + w){
